main: Display password verifier instead of the password.
Move the field where the password is rendered off the visible area of the
display. Owing to a quirk of Chrome, you can't copy from a text field that is
invisible or has zero size. However, you can copy from one that is 1x1 pixels
in size and off the top left corner of the viewport.