Fix handling of the -punct flag.
This flag is supposed to override what is in the config file, but the way it
was hooked up it only affected the default configuration before reading from
the file. Moreover, setting the default was not enough, since a site may
explicitly set "punct":false.

Fix this by separating the flag from the config, and plugging in the value at
the point of key generation.