Rework how static configurations are embedded.
Replace my custom staticfile package with the built-in go:embed.

Add a default-empty embed.FS, and update the go:generate rule to add a file to
the package that populates that FS. This is a little trickier than it seems,
because the //go:embed comment has to go right before the FS it targets, but we
don't want to edit the existing package files.

So instead, the generated code embeds the config into a placeholder FS, and
then copies that value into the "real" one during init. We'll pretend I just
figured that out right away rather than spending like an hour dicking around
with more complicated solutions.