kflib: add HashedChars, update tests
This is a step toward replacing the old hashpass algorithm.  Instead of an
ad-hoc HMAC, use HKDF to derive key material for the given site seed, and
consume entropy in a more conservative manner.

I can't replace all the existing usage yet, because this gives different
results than the original algorithm.