cmd/kf::server: add the ability to "lock" the UI
Clicking the Keyfish icon marks the UI as "locked", and it will not allow
searches or other operations until the correct PIN is provided.

This is not really a security feature, it's just a way to slow down someone who
happens on a browser window with the tab already open.