cmd/kf: add a private --kf.pfile flag
When this flag is set, it identifies a file from which the passphrase to open
the database will be read. This allows the caller to set up a pipe to send in
the passphrase at startup, so a server can be started in the background without
having to block at the terminal first.